# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Rumi Seismic Service API",
        "description": "REST API of the Seismic Correlation bounded context: seismic events reported by the IGP (US13), risk indexes of buildings (US11) and their affected zones (US12). The service consumes the SensorReadingRecorded event over RabbitMQ and correlates it with the IGP events. Errors use RFC 7807 ProblemDetail.",
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8083",
            "description": "Generated server url"
        }
    ],
    "tags": [
        {
            "name": "Risk indexes",
            "description": "Risk of a building after a seismic event and its affected zones (US11, US12)"
        },
        {
            "name": "Seismic events",
            "description": "Earthquakes reported by the IGP (US13)"
        }
    ],
    "paths": {
        "/api/v1/seismic-events": {
            "get": {
                "tags": [
                    "Seismic events"
                ],
                "summary": "List the stored seismic events",
                "description": "US13. Returns the events stored by Rumi, newest first. from and to are optional ISO-8601 instants, both inclusive; without them every stored event is returned. This endpoint reads the database only and does not call the IGP.",
                "operationId": "listEvents",
                "parameters": [
                    {
                        "name": "from",
                        "in": "query",
                        "description": "Earliest occurrence time, inclusive (ISO-8601 instant, optional)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        },
                        "example": "2026-10-01T00:00:00Z"
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "Latest occurrence time, inclusive (ISO-8601 instant, optional)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        },
                        "example": "2026-10-07T00:00:00Z"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Stored events, newest first (empty list if there are none in the range)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SeismicEventResponse"
                                    }
                                },
                                "examples": {
                                    "seismicEvents": {
                                        "summary": "Two stored events",
                                        "description": "seismicEvents",
                                        "value": [
                                            {
                                                "id": "1b2c3d4e-5f60-4718-8293-a4b5c6d7e8f9",
                                                "occurredAt": "2026-10-06T15:29:41Z",
                                                "magnitude": {
                                                    "value": 5.8,
                                                    "scale": "Mw"
                                                },
                                                "epicenter": {
                                                    "latitude": -12.05,
                                                    "longitude": -77.12,
                                                    "description": "Callao, 35 km SW"
                                                },
                                                "source": "IGP"
                                            },
                                            {
                                                "id": "2c3d4e5f-6071-4829-93a4-b5c6d7e8f9a0",
                                                "occurredAt": "2026-10-03T22:10:05Z",
                                                "magnitude": {
                                                    "value": 4.2,
                                                    "scale": "ML"
                                                },
                                                "epicenter": {
                                                    "latitude": -14.07,
                                                    "longitude": -75.73,
                                                    "description": "Ica, 20 km W"
                                                },
                                                "source": "IGP"
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "from or to is not an ISO-8601 instant, or from is after to",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                },
                                "examples": {
                                    "malformedFrom": {
                                        "summary": "from is not an ISO-8601 instant",
                                        "description": "malformedFrom",
                                        "value": {
                                            "type": "about:blank",
                                            "title": "Bad Request",
                                            "status": 400,
                                            "detail": "from must be an ISO-8601 instant, for example 2026-10-06T15:30:00Z",
                                            "instance": "/api/v1/seismic-events"
                                        }
                                    },
                                    "reversedRange": {
                                        "summary": "from is after to",
                                        "description": "reversedRange",
                                        "value": {
                                            "type": "about:blank",
                                            "title": "Bad Request",
                                            "status": 400,
                                            "detail": "from must not be after to",
                                            "instance": "/api/v1/seismic-events"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/seismic-events/latest": {
            "get": {
                "tags": [
                    "Seismic events"
                ],
                "summary": "Get the latest seismic event reported by the IGP",
                "description": "US13. Asks the IGP feed for its latest event through a circuit breaker and stores it if it is new (same occurrence time, magnitude and epicenter means the same event). If the IGP fails, times out or its circuit is open, the latest stored event is returned instead; 503 only when there is no stored event either.",
                "operationId": "getLatestEvent",
                "responses": {
                    "200": {
                        "description": "Latest event, from the IGP or, if it is unavailable, the latest stored one",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SeismicEventResponse"
                                },
                                "examples": {
                                    "latestEvent": {
                                        "summary": "Earthquake off Callao",
                                        "description": "latestEvent",
                                        "value": {
                                            "id": "1b2c3d4e-5f60-4718-8293-a4b5c6d7e8f9",
                                            "occurredAt": "2026-10-06T15:29:41Z",
                                            "magnitude": {
                                                "value": 5.8,
                                                "scale": "Mw"
                                            },
                                            "epicenter": {
                                                "latitude": -12.05,
                                                "longitude": -77.12,
                                                "description": "Callao, 35 km SW"
                                            },
                                            "source": "IGP"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "The IGP is unavailable and there is no stored event to fall back to",
                        "headers": {
                            "Retry-After": {
                                "description": "Seconds to wait before retrying",
                                "style": "simple",
                                "schema": {
                                    "type": "integer",
                                    "example": 60
                                }
                            }
                        },
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                },
                                "examples": {
                                    "igpUnavailable": {
                                        "summary": "IGP down and no stored event",
                                        "description": "igpUnavailable",
                                        "value": {
                                            "type": "about:blank",
                                            "title": "Service Unavailable",
                                            "status": 503,
                                            "detail": "The IGP feed is unavailable and no seismic event is stored yet",
                                            "instance": "/api/v1/seismic-events/latest"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/risk-indexes/{riskIndexId}/affected-zones": {
            "get": {
                "tags": [
                    "Risk indexes"
                ],
                "summary": "List the affected zones of a risk index",
                "description": "US12. Returns the zones of the building affected by the seismic event, ordered by severity (rank 1 is the most affected). Zones are only reported when the risk level is MEDIUM or higher, so a LOW risk index returns an empty list.",
                "operationId": "getAffectedZones",
                "parameters": [
                    {
                        "name": "riskIndexId",
                        "in": "path",
                        "description": "Risk index whose affected zones are listed",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "example": "c3a1e8d2-6b7f-4d09-9a21-5e8f7b6c4d3a"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Affected zones ordered by severity rank (empty list below MEDIUM)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AffectedZoneResponse"
                                    }
                                },
                                "examples": {
                                    "affectedZones": {
                                        "summary": "Two affected zones of a HIGH risk index",
                                        "description": "affectedZones",
                                        "value": [
                                            {
                                                "zone": "FLOOR-3-NORTH",
                                                "severityRank": 1
                                            },
                                            {
                                                "zone": "FLOOR-2-NORTH",
                                                "severityRank": 2
                                            }
                                        ]
                                    },
                                    "lowRisk": {
                                        "summary": "LOW risk index, no affected zones",
                                        "description": "lowRisk",
                                        "value": []
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "riskIndexId is not a UUID",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                },
                                "examples": {
                                    "malformedRiskIndexId": {
                                        "summary": "riskIndexId is not a UUID",
                                        "description": "malformedRiskIndexId",
                                        "value": {
                                            "type": "about:blank",
                                            "title": "Bad Request",
                                            "status": 400,
                                            "detail": "riskIndexId must be a valid UUID",
                                            "instance": "/api/v1/risk-indexes/latest-risk/affected-zones"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The risk index does not exist",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                },
                                "examples": {
                                    "unknownRiskIndex": {
                                        "summary": "Unknown risk index",
                                        "description": "unknownRiskIndex",
                                        "value": {
                                            "type": "about:blank",
                                            "title": "Not Found",
                                            "status": 404,
                                            "detail": "Risk index 6d5c4b3a-2f1e-4d0c-9b8a-7f6e5d4c3b2a does not exist",
                                            "instance": "/api/v1/risk-indexes/6d5c4b3a-2f1e-4d0c-9b8a-7f6e5d4c3b2a/affected-zones"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/buildings/{buildingId}/risk-indexes/latest": {
            "get": {
                "tags": [
                    "Risk indexes"
                ],
                "summary": "Get the latest risk index of a building",
                "description": "US11. Returns the most recently calculated risk index of the building, with the correlation it comes from and the model that calculated it. A risk index is created when a seismic event reported by the IGP matches the sensor readings of the building.",
                "operationId": "getLatestRiskIndex",
                "parameters": [
                    {
                        "name": "buildingId",
                        "in": "path",
                        "description": "Building whose risk index is queried",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "example": "7a9b3c1d-2e4f-4b6a-8c0d-1e2f3a4b5c6d"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Latest risk index of the building",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RiskIndexResponse"
                                },
                                "examples": {
                                    "latestRiskIndex": {
                                        "summary": "HIGH risk after the Callao earthquake",
                                        "description": "latestRiskIndex",
                                        "value": {
                                            "id": "c3a1e8d2-6b7f-4d09-9a21-5e8f7b6c4d3a",
                                            "buildingId": "7a9b3c1d-2e4f-4b6a-8c0d-1e2f3a4b5c6d",
                                            "correlatedEventId": "9e8d7c6b-5a49-4382-b1a0-f9e8d7c6b5a4",
                                            "level": "HIGH",
                                            "calculatedAt": "2026-10-06T15:32:10Z",
                                            "modelVersion": "rule-based-v1"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "buildingId is not a UUID",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                },
                                "examples": {
                                    "malformedBuildingId": {
                                        "summary": "buildingId is not a UUID",
                                        "description": "malformedBuildingId",
                                        "value": {
                                            "type": "about:blank",
                                            "title": "Bad Request",
                                            "status": 400,
                                            "detail": "buildingId must be a valid UUID",
                                            "instance": "/api/v1/buildings/building-7/risk-indexes/latest"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The building has no risk index",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                },
                                "examples": {
                                    "noRiskIndex": {
                                        "summary": "No seismic event has been correlated with the building",
                                        "description": "noRiskIndex",
                                        "value": {
                                            "type": "about:blank",
                                            "title": "Not Found",
                                            "status": 404,
                                            "detail": "Building 0f1e2d3c-4b5a-4968-8776-5a4b3c2d1e0f has no risk index",
                                            "instance": "/api/v1/buildings/0f1e2d3c-4b5a-4968-8776-5a4b3c2d1e0f/risk-indexes/latest"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "EpicenterResponse": {
                "type": "object",
                "description": "Epicenter of the event",
                "properties": {
                    "latitude": {
                        "type": "number",
                        "format": "double",
                        "description": "Latitude in degrees",
                        "example": -12.05
                    },
                    "longitude": {
                        "type": "number",
                        "format": "double",
                        "description": "Longitude in degrees",
                        "example": -77.12
                    },
                    "description": {
                        "type": "string",
                        "description": "Reference of the epicenter",
                        "example": "Callao, 35 km SW"
                    }
                }
            },
            "MagnitudeResponse": {
                "type": "object",
                "description": "Magnitude of the event",
                "properties": {
                    "value": {
                        "type": "number",
                        "format": "double",
                        "description": "Magnitude value",
                        "example": 5.8
                    },
                    "scale": {
                        "type": "string",
                        "description": "Magnitude scale",
                        "example": "Mw"
                    }
                }
            },
            "SeismicEventResponse": {
                "type": "object",
                "description": "Seismic event reported by the IGP and stored by Rumi",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "description": "Id of the stored event",
                        "example": "1b2c3d4e-5f60-4718-8293-a4b5c6d7e8f9"
                    },
                    "occurredAt": {
                        "type": "string",
                        "format": "date-time",
                        "description": "When the earthquake occurred",
                        "example": "2026-10-06T15:29:41Z"
                    },
                    "magnitude": {
                        "$ref": "#/components/schemas/MagnitudeResponse"
                    },
                    "epicenter": {
                        "$ref": "#/components/schemas/EpicenterResponse"
                    },
                    "source": {
                        "type": "string",
                        "description": "Agency that reported the event",
                        "example": "IGP"
                    }
                }
            },
            "ProblemDetail": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string",
                        "format": "uri"
                    },
                    "title": {
                        "type": "string"
                    },
                    "status": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "detail": {
                        "type": "string"
                    },
                    "instance": {
                        "type": "string",
                        "format": "uri"
                    },
                    "properties": {
                        "type": "object",
                        "additionalProperties": {}
                    }
                }
            },
            "AffectedZoneResponse": {
                "type": "object",
                "description": "Zone of the building affected by the seismic event",
                "properties": {
                    "zone": {
                        "type": "string",
                        "description": "Zone of the building",
                        "example": "FLOOR-3-NORTH"
                    },
                    "severityRank": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Severity rank, 1 is the most affected zone",
                        "example": 1
                    }
                }
            },
            "RiskIndexResponse": {
                "type": "object",
                "description": "Risk index of a building after a seismic event",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "description": "Risk index id",
                        "example": "c3a1e8d2-6b7f-4d09-9a21-5e8f7b6c4d3a"
                    },
                    "buildingId": {
                        "type": "string",
                        "format": "uuid",
                        "description": "Building the risk index belongs to",
                        "example": "7a9b3c1d-2e4f-4b6a-8c0d-1e2f3a4b5c6d"
                    },
                    "correlatedEventId": {
                        "type": "string",
                        "format": "uuid",
                        "description": "Correlation of the seismic event with the building",
                        "example": "9e8d7c6b-5a49-4382-b1a0-f9e8d7c6b5a4"
                    },
                    "level": {
                        "type": "string",
                        "description": "Risk level",
                        "enum": [
                            "LOW",
                            "MEDIUM",
                            "HIGH",
                            "CRITICAL"
                        ],
                        "example": "HIGH"
                    },
                    "calculatedAt": {
                        "type": "string",
                        "format": "date-time",
                        "description": "When the risk index was calculated",
                        "example": "2026-10-06T15:32:10Z"
                    },
                    "modelVersion": {
                        "type": "string",
                        "description": "Model that calculated the risk index",
                        "example": "rule-based-v1"
                    }
                }
            }
        }
    }
}